+++ src/test/java/org/scion/jpan/api/ScionTest.java
@@ -317,4 +317,22 @@ public class ScionTest {
       MockNetwork.stopTiny();
     }
   }
+
+  /** See Issue #72: https://github.com/scionproto-contrib/jpan/issues/72 */
+  @Test
+  void defaultService_bootstrapTopoFile_ScionProto_11() {
+    long dstIA = ScionUtil.parseIA("1-ff00:0:112");
+    InetSocketAddress dstAddress = new InetSocketAddress("::1", 12345);
+    MockNetwork.startTiny(MockNetwork.Mode.BOOTSTRAP);
+    try {
+      System.setProperty(
+          Constants.PROPERTY_BOOTSTRAP_TOPO_FILE, "topologies/topology-scionproto-0.11.json");
+      ScionService service = Scion.defaultService();
+      RequestPath path = service.getPaths(dstIA, dstAddress).get(0);
+      assertNotNull(path);
+      assertEquals(0, MockDaemon.getAndResetCallCount()); // Daemon is not used!
+    } finally {
+      MockNetwork.stopTiny();
+    }
+  }
 }